factor out ex_any.hpp and work_item.hpp
